Course Crud

+++ src/main/java/com/sportifyindia/app/service/CourseService.java
@@ -1,10 +1,16 @@
 package com.sportifyindia.app.service;
 
 import com.sportifyindia.app.domain.Course;
+import com.sportifyindia.app.domain.Facility;
+import com.sportifyindia.app.domain.User;
 import com.sportifyindia.app.repository.CourseRepository;
+import com.sportifyindia.app.repository.FacilityRepository;
+import com.sportifyindia.app.repository.UserRepository;
 import com.sportifyindia.app.repository.search.CourseSearchRepository;
+import com.sportifyindia.app.security.SecurityUtils;
 import com.sportifyindia.app.service.dto.CourseDTO;
 import com.sportifyindia.app.service.mapper.CourseMapper;
+import java.util.List;
 import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,7 +20,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
- * Service Implementation for managing {@link com.sportifyindia.app.domain.Course}.
+ * Service Implementation for managing {@link Course}.
  */
 @Service
 @Transactional
@@ -23,14 +29,22 @@ public class CourseService {
     private final Logger log = LoggerFactory.getLogger(CourseService.class);
 
     private final CourseRepository courseRepository;
-
     private final CourseMapper courseMapper;
-
+    private final FacilityRepository facilityRepository;
+    private final UserRepository userRepository;
     private final CourseSearchRepository courseSearchRepository;
 
-    public CourseService(CourseRepository courseRepository, CourseMapper courseMapper, CourseSearchRepository courseSearchRepository) {
+    public CourseService(
+        CourseRepository courseRepository,
+        CourseMapper courseMapper,
+        FacilityRepository facilityRepository,
+        UserRepository userRepository,
+        CourseSearchRepository courseSearchRepository
+    ) {
         this.courseRepository = courseRepository;
         this.courseMapper = courseMapper;
+        this.facilityRepository = facilityRepository;
+        this.userRepository = userRepository;
         this.courseSearchRepository = courseSearchRepository;
     }
 
@@ -42,6 +56,7 @@ public class CourseService {
      */
     public CourseDTO save(CourseDTO courseDTO) {
         log.debug("Request to save Course : {}", courseDTO);
+        checkUserHasAccess(courseDTO.getFacilityId());
         Course course = courseMapper.toEntity(courseDTO);
         course = courseRepository.save(course);
         CourseDTO result = courseMapper.toDto(course);
@@ -52,11 +67,12 @@ public class CourseService {
     /**
      * Update a course.
      *
-     * @param courseDTO the entity to save.
+     * @param courseDTO the entity to update.
      * @return the persisted entity.
      */
     public CourseDTO update(CourseDTO courseDTO) {
         log.debug("Request to update Course : {}", courseDTO);
+        checkUserHasAccess(courseDTO.getFacilityId());
         Course course = courseMapper.toEntity(courseDTO);
         course = courseRepository.save(course);
         CourseDTO result = courseMapper.toDto(course);
@@ -72,12 +88,11 @@ public class CourseService {
      */
     public Optional<CourseDTO> partialUpdate(CourseDTO courseDTO) {
         log.debug("Request to partially update Course : {}", courseDTO);
-
+        checkUserHasAccess(courseDTO.getFacilityId());
         return courseRepository
             .findById(courseDTO.getId())
             .map(existingCourse -> {
                 courseMapper.partialUpdate(existingCourse, courseDTO);
-
                 return existingCourse;
             })
             .map(courseRepository::save)
@@ -101,6 +116,15 @@ public class CourseService {
     }
 
     /**
+     * Get all the courses with eager load of many-to-many relationships.
+     *
+     * @return the list of entities.
+     */
+    public Page<CourseDTO> findAllWithEagerRelationships(Pageable pageable) {
+        return courseRepository.findAllWithEagerRelationships(pageable).map(courseMapper::toDto);
+    }
+
+    /**
      * Get one course by id.
      *
      * @param id the id of the entity.
@@ -109,7 +133,7 @@ public class CourseService {
     @Transactional(readOnly = true)
     public Optional<CourseDTO> findOne(Long id) {
         log.debug("Request to get Course : {}", id);
-        return courseRepository.findById(id).map(courseMapper::toDto);
+        return courseRepository.findOneWithEagerRelationships(id).map(courseMapper::toDto);
     }
 
     /**
@@ -119,8 +143,12 @@ public class CourseService {
      */
     public void delete(Long id) {
         log.debug("Request to delete Course : {}", id);
-        courseRepository.deleteById(id);
-        courseSearchRepository.deleteFromIndexById(id);
+        Optional<Course> course = courseRepository.findById(id);
+        if (course.isPresent()) {
+            checkUserHasAccess(course.get().getFacility().getId());
+            courseRepository.deleteById(id);
+            courseSearchRepository.deleteFromIndexById(id);
+        }
     }
 
     /**
@@ -134,5 +162,36 @@ public class CourseService {
     public Page<CourseDTO> search(String query, Pageable pageable) {
         log.debug("Request to search for a page of Courses for query {}", query);
         return courseSearchRepository.search(query, pageable).map(courseMapper::toDto);
+    }
+
+    /**
+     * Get all the courses by facility ID.
+     *
+     * @param facilityId the ID of the facility.
+     * @param pageable the pagination information.
+     * @return the list of entities.
+     */
+    @Transactional(readOnly = true)
+    public Page<CourseDTO> findAllByFacilityId(Long facilityId, Pageable pageable) {
+        log.debug("Request to get all Courses for facility : {}", facilityId);
+        return courseRepository.findByFacilityId(facilityId, pageable).map(courseMapper::toDto);
+    }
+
+    private void checkUserHasAccess(Long facilityId) {
+        Optional<User> user = userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin().orElse(""));
+        if (user.isPresent()) {
+            Optional<Facility> facility = facilityRepository.findById(facilityId);
+            if (facility.isPresent()) {
+                if (
+                    !user
+                        .get()
+                        .getAuthorities()
+                        .stream()
+                        .anyMatch(auth -> auth.getName().equals("ROLE_OWNER") || auth.getName().equals("ROLE_ADMIN"))
+                ) {
+                    throw new RuntimeException("User does not have permission to perform this action");
+                }
+            }
+        }
     }
 }
